% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veri_setleri.r
\docType{data}
\name{secim151101g}
\alias{secim151101g}
\title{1 Kasım 2015 Genel Seçimleri Sandık Bazında Seçmen ve Oy Sayıları}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 177603 rows and 26 columns.}
\usage{
secim151101g
}
\arguments{
\item{il}{İlgili sandığın seçim ilini ifade eder.}

\item{bolge}{Sadece Ankara (2), İstanbul (3) ve İzmir (2) seçim bölgelerine ayrılmışlardır. Diğer bütün iller için bölge numarası 1'dir.}

\item{ilce}{İlgili sandığın ilçesini ifade eder. Satırlarda illerin toplam istatistiklerini belirten değer İli ile ifade edilmiştir, diğer bütün değerler ilçe isimleridir.}

\item{cevre}{İlgili sandığın seçim çevresi isimlerini içerir. Bu sütundaki özel değerler Genel, İl-İlçe Genel ve Köy-Belde Genel ifadelerinden oluşmaktadır. Özel değerler sırasıyla ilin toplam sandık ve oy sayılarını, ilçelerinin ve merkezinin toplam sandık ve oy sayılarını ve köylerinin ve beldelerinin toplam sandık ve oy sayılarını belirtir.}

\item{cevre_turu}{İlgili sandığın bulunduğu seçim çevresinin türünü içerir. Seçim çevreleri Köy, Belde, Mahalle veya Cezaevi olabilirler. Bu sütundaki Toplam değeri ilgili satırın bir sandık verisi içermediğini, birden çok sandığın toplamının verilerini içerir. Toplamın neleri kapsadığını aynı satırın kendisinden önceki sütunlarında verilen bilgilerde bulabilirsiniz.}

\item{sandik}{Sandık numarasını veya toplam değerdeki sandık sayısını belirtir. Eğer aynı satırda `cevre_turu` değeri Toplam ise toplam sandık sayısını, başka bir değer ise sandık numarasını verir.}

\item{kayitli_secmen}{İlgili sandıkta (eğer `cevre_turu` Toplam ise ilgili bölgede) kayıtlı toplam seçmen sayısını verir.}

\item{oy_kullanan}{İlgili sandıkta (eğer `cevre_turu` Toplam ise ilgili bölgede) oy kullanan toplam kişi sayısını verir. Bazı sandıklarda oy kullanan kişi sayısı kayıtlı seçmen sayısından fazla olabilir. Bunun sebepleri arasında görevlilerin (ör. sandık görevlisi, polis memuru) tutanakla kendi kayıtlı sandıklarının yerine başka sandıklarda oy atabilmeleri ve kayıt hataları bulunmaktadır.}

\item{gecerli_oy}{İlgili sandıktaki (eğer `cevre_turu` Toplam ise ilgili bölgede) geçerli oy sayısını verir. Oylar çeşitli sebeplerle geçerli sayılmayabilmektedir. (Ör. Boş oy, çift mühür, oyun nereye basıldığının anlaşılamaması.)}

\item{ak_parti}{Adalet ve Kalkınma Partisi (AK PARTİ) 14 Ağustos 2001 tarihinde kurulmuştur. Merkez sağ partisidir. İdeolojileri sosyal muhafazakarlık, muhafazakar demokrasi, ekonomik liberalizm ve yeni Osmanlıcılıktır. 2002 genel seçimlerinden Kasım 2015 genel seçimlerine kadar en yüksek oyu alıp iktidar partisi olmuşlardır. (Kaynak: Vikipedi)}

\item{chp}{Cumhuriyet Halk Partisi (CHP) 9 Eylül 1923'te Mustafa Kemal Atatürk tarafından kurulmuştur. Merkez sol partisidir. İdeolojileri Atatürkçülük, sosyal demokrasi ve demokratik soldur. 2002 genel seçimlerinden Kasım 2015 genel seçimlerine kadar en yüksek ikinci oyu alıp ana muhalefet partisi olmuşlardır. (Kaynak: Vikipedi)}

\item{mhp}{Milliyetçi Hareket Partisi (MHP) 9 Şubat 1969 tarihinde kurulmuştur. Sağ partidir. İdeolojileri Türk milliyetçiliği, ülkücülük ve dokuz ışıktır. Kasım 2015 genel seçimlerinde en yüksek üçüncü oyu alıp muhalefet olmuşlardır. (Kaynak: Vikipedi)}

\item{hdp}{Halkların Demokratik Partisi (HDP) 27 Ekim 2013 tarihinde kurulmuştur. Sol partidir. İdeolojileri siyasi çoğulculuk, sosyalizm, radikal demokrasi, demokratik sosyalizm, feminizm ve doğrudan demokrasidir. Kasım 2015 genel seçimlerinde en yüksek dördüncü oyu alıp muhalefet olmuşlardır. (Kaynak: Vikipedi)}

\item{bbp}{Büyük Birlik Partisi (BBP)}

\item{btp}{Bağımsız Türkiye Partisi (BTP)}

\item{dp}{Demokrat Parti (DP)}

\item{dsp}{Demokratik Sol Parti (DSP)}

\item{dyp}{Doğru Yol Partisi (DYP)}

\item{hak_par}{Hak ve Özgürlükler Partisi (HAK-PAR)}

\item{hkp}{Halkın Kurtuluş Partisi (HKP)}

\item{ldp}{Liberal Demokrat Parti (LDP)}

\item{millet_partisi}{Millet Partisi}

\item{saadet_partisi}{Saadet Partisi}

\item{vatan_partisi}{Vatan Partisi}

\item{bagimsiz}{Bağımsız adaylar.}
}
\description{
Bu veri setinde 1 Kasım 2015 Milletvekili Genel Seçimi'nin sandık bazında seçmen bilgileri ve oy sayıları bulunmaktadır. Veri setinin bazı satırlarında toplam bilgileri verilmiştir. Her sütunun açıklamasında toplam bilgilerinin nasıl ifade edildiği belirtilmektedir.
}
\keyword{datasets}

